// src/application/dto/farm/update-farm.output.dto.ts
export type UpdateFarmOutputDto = {
  id: string;
  name: string;
  city: string;
  state: string;
  totalAreaHectares: number;
  cultivableAreaHectares: number;
  vegetationAreaHectares: number;
  crops: {
    id: string;
    name: string;
  }[];
  farmer: {
    id: string;
    name: string;
    cpfCnpj: string;
  };
};